# pkg/sdk/openapi/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "跨境电商运营 Agent REST API，由 swaggo/swag 从 handler 注解生成。",
        "title": "E-commerce Operations Agent API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/prompts": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "prompts"
                ],
                "summary": "列出 Prompt",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "additionalProperties": true
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "prompts"
                ],
                "summary": "创建 Prompt(含 v1)",
                "parameters": [
                    {
                        "description": "Prompt",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/prompt.CreatePromptRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "prompt.CreatePromptRequest": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string"
                },
                "created_by": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "template": {
                    "type": "string"
                },
                "variables_schema": {
                    "type": "string"
                },
                "workspace_id": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}